requirements.txt
asttokens==3.0.0
blinker==1.9.0
click==8.2.1
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kafka-python==2.2.15
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.3.1
packaging==25.0
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg==3.2.9
psycopg-binary==3.2.9
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.9
Pygments==2.19.2
pyspark==4.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.0.0
schedule==1.2.2
scikit-learn==1.7.0
scipy==1.16.0
six==1.17.0
SQLAlchemy==2.0.41
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
wcwidth==0.2.13
Werkzeug==3.1.3
